+++ core/src/main/java/com/yahoo/tagchowder/CommandLine.java
@@ -21,7 +21,7 @@
 
 // The TagChowder command line UI
 
-package com.lafaspot.tagchowder;
+package com.yahoo.tagchowder;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -37,7 +37,7 @@ import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
 import org.xml.sax.ext.LexicalHandler;
 
-import com.lafaspot.tagchowder.templates.HTMLSchema;
+import com.yahoo.tagchowder.templates.HTMLSchema;
 
 /**
  * The stand-alone TagChowder program.
