+++ core/src/main/java/com/yahoo/tagchowder/jaxp/SAX1ParserAdapter.java
@@ -19,7 +19,7 @@
  * Changes to the original project are Copyright 2019 Oath Inc.
  */
 
-package com.lafaspot.tagchowder.jaxp;
+package com.yahoo.tagchowder.jaxp;
 
 import java.io.IOException;
