+++ core/src/main/java/com/yahoo/tagchowder/templates/package-info.java
@@ -16,6 +16,6 @@
  */
 
 /**
- * Package info of com.lafaspot.tagchowder.jaxp.
+ * Package info of com.yahoo.tagchowder.templates.
  */
-package com.lafaspot.tagchowder.jaxp;
+package com.yahoo.tagchowder.templates;
