+++ core/src/test/java/com/yahoo/tagchowder/ParserTest.java
@@ -19,7 +19,7 @@
  * Changes to the original project are Copyright 2019 Oath Inc.
  */
 
-package com.lafaspot.tagchowder;
+package com.yahoo.tagchowder;
 
 import java.io.BufferedReader;
 import java.io.IOException;
