+++ pom.xml
@@ -3,21 +3,21 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.github.lafa.tagchowder</groupId>
+    <groupId>com.yahoo.tagchowder</groupId>
     <artifactId>tagchowder</artifactId>
     <packaging>pom</packaging>
-    <version>2.0.3</version>
+    <version>2.0.4</version>
     <name>${project.artifactId}</name>
-    <url>https://github.com/lafaspot/tagchowder</url>
+    <url>https://github.com/yahoo/tagchowder</url>
     <description>Parent POM file for tagchowder project</description>
     <prerequisites>
         <maven>3.0</maven>
     </prerequisites>
 
     <scm>
-        <developerConnection>scm:git:https://github.com/lafaspot/tagchowder.git</developerConnection>
-        <connection>scm:git:https://github.com/lafaspot/tagchowder.git</connection>
-        <url>https://github.com/lafaspot/tagchowder.git</url>
+        <developerConnection>scm:git:https://github.com/yahoo/tagchowder.git</developerConnection>
+        <connection>scm:git:https://github.com/yahoo/tagchowder.git</connection>
+        <url>https://github.com/yahoo/tagchowder.git</url>
         <tag>tagchowder-1.0</tag>
     </scm>
     <developers>
@@ -35,7 +35,7 @@
 
     <issueManagement>
         <system>GitHub</system>
-        <url>https://github.com/lafaspot/tagchowder/issues</url>
+        <url>https://github.com/yahoo/tagchowder/issues</url>
     </issueManagement>
 
     <licenses>
